+++ src/main/java/com/kalsym/order/service/enums/RefundStatus.java
@@ -1,0 +1,12 @@
+
+package com.kalsym.order.service.enums;
+
+/**
+ *
+ * @author FaisalHayatJadoon
+ */
+public enum RefundStatus {
+    PENDING,
+    COMPLETED,
+    FAILED
+}