+++ src/main/java/com/kalsym/order/service/enums/RefundType.java
@@ -1,0 +1,12 @@
+
+package com.kalsym.order.service.enums;
+
+/**
+ *
+ * @author FaisalHayatJadoon
+ */
+public enum RefundType {
+    ORDER_CANCELLED,
+    ITEM_CANCELLED,
+    ITEM_REVISED
+}