+++ src/main/java/com/kalsym/order/service/model/PaymentOrder.java
@@ -1,0 +1,49 @@
+package com.kalsym.order.service.model;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import java.util.Date;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+import org.hibernate.annotations.GenericGenerator;
+import java.util.List;
+import javax.persistence.CascadeType;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.Temporal;
+import javax.persistence.TemporalType;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import com.kalsym.order.service.enums.RefundType;
+import com.kalsym.order.service.enums.RefundStatus;
+
+/**
+ *
+ * @author 7cu
+ */
+@Entity
+@Getter
+@Setter
+@ToString
+@Table(name = "payment_orders")
+public class PaymentOrder {
+
+    @Id
+    @GeneratedValue(generator = "uuid2")
+    @GenericGenerator(name = "uuid2", strategy = "org.hibernate.id.UUIDGenerator")
+    private String id;
+
+    private String clientTransactionId;   
+    private String paymentChannel;
+    private String status;
+   
+}